# BatchCompleter: batch all ops, not just completed