@bgentry 2024-09-25
@brandur alright, this is ready to go, please take a look if you can. Also let me know if you think we should nuke the inline or async completers altogether at this point (along with their DB queries) or if you think we should keep them around for now.