@bgentry 2024-09-25
The `InlineCompleter` is still used from some tests. Removing just the `AsyncCompleter` won't save many LoC because the underlying query is shared with `InlineCompleter` anyway, so I think it makes sense to just leave those around for now.

Going to move forward on shipping the rest of this to unblock upcoming stuff.